+++ src/main/java/org/reduct/model/bucket/Bucket.java
@@ -3,28 +3,29 @@ package org.reduct.model.bucket;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.*;
-import org.apache.commons.lang3.SerializationUtils;
+import org.apache.commons.lang3.ArrayUtils;
 import org.reduct.client.ReductClient;
 import org.reduct.common.BucketURL;
-import org.reduct.common.EntryURL;
+import org.reduct.common.RecordURL;
 import org.reduct.common.exception.ReductException;
-import org.reduct.common.exception.ReductSDKException;
-import org.reduct.model.entry.Entry;
 import org.reduct.model.mapper.BucketMapper;
+import org.reduct.model.record.QueryId;
+import org.reduct.model.record.Record;
 import org.reduct.utils.JsonUtils;
 import org.reduct.utils.Strings;
-import org.reduct.utils.http.HttpHeaders;
+import org.reduct.utils.http.Queries;
 
-import java.io.Serializable;
 import java.math.BigInteger;
 import java.net.URI;
 import java.net.http.HttpRequest;
 import java.net.http.HttpResponse;
-import java.util.List;
-import java.util.Map;
-import java.util.Objects;
+import java.nio.ByteBuffer;
+import java.time.Instant;
+import java.util.*;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
-import static org.reduct.utils.http.Query.TIME_STAMP;
+import static org.reduct.utils.http.HttpHeaders.*;
 
 @NoArgsConstructor
 @EqualsAndHashCode
@@ -33,6 +34,13 @@ import static org.reduct.utils.http.Query.TIME_STAMP;
 @Getter
 @Setter
 public class Bucket {
+
+    private static final String TS = "ts";
+    public static final String BUCKET_NAME_CANNOT_BE_NULL_OR_EMPTY = "Bucket name cannot be null or empty.";
+    public static final String X_REDUCT_TIME_IS_NOT_SUCH_LONG_FORMAT = "Received from server x-reduct-time is not such Long format, or empty.";
+    public static final String CONTENT_TYPE_IS_NOT_SET_IN_THE_RECORD = "The Content-Type is not set in the record.";
+    public static final String CONTENT_LENGTH_IS_NOT_SET_IN_THE_RECORD = "The Content-Length is not set in the record.";
+
     public Bucket(String name, ReductClient reductClient) {
         this.name = name;
         this.reductClient = reductClient;
@@ -86,28 +94,26 @@ public class Bucket {
     private List<EntryInfo> entryInfos;
 
     /**
+     * Get information about a bucket
      * The method returns the current settings, stats, and entry list of the bucket in JSON format. If authentication is enabled, the method needs a valid API token.
      * @return Returns this Bucket object with updated fields
      */
-    public Bucket read() throws ReductException, ReductSDKException, IllegalArgumentException {
+    public Bucket read() throws ReductException, IllegalArgumentException {
         if (name == null || name.isBlank()) {
-            throw new IllegalArgumentException("Bucket name cannot be null or empty");
+            throw new IllegalArgumentException(BUCKET_NAME_CANNOT_BE_NULL_OR_EMPTY);
         }
         String createBucketPath = BucketURL.GET_BUCKET.getUrl().formatted(name);
         HttpRequest.Builder builder = HttpRequest.newBuilder()
                 .uri(URI.create("%s/%s".formatted(reductClient.getServerProperties().getBaseUrl(), createBucketPath)))
                 .GET();
         HttpResponse<String> httpResponse = reductClient.send(builder, HttpResponse.BodyHandlers.ofString());
-        Bucket read = JsonUtils.parseObject(httpResponse.body(), Bucket.class);
-        BucketMapper.INSTANCE.copy(this, read);
+        BucketMapper.INSTANCE.copy(this, JsonUtils.parseObject(httpResponse.body(), Bucket.class));
         return this;
     }
 
     /**
-     * Create a new bucket with the name and settings.
-     * NOTE: If, authentication is enabled on the server, an access token with full access must be provided
-     * to create a new bucket.
-     * @return This bucket
+     * To update settings of a bucket, the request should have a JSON document with all the settings.
+     * @param bucketSettings
      * @throws ReductException          If, unable to create the bucket. The instance of the exception holds
      *                                  the error message returned in the x-reduct-error header and the
      *                                  status code to indicate the failure.
@@ -117,101 +123,251 @@ public class Bucket {
      *                                  409 -> Bucket with this name already exists.
      *                                  422 -> Invalid request.
      *                                  500 -> Internal server error.
-     * @throws ReductSDKException       If, any client side error occur.
+     * @throws ReductException       If, any client side error occur.
      * @throws IllegalArgumentException If, the bucket name is null or empty.
      */
-    public Bucket write() throws ReductException, ReductSDKException, IllegalArgumentException {
+    @JsonIgnore
+    public void setSettings(BucketSettings bucketSettings) throws ReductException, IllegalArgumentException {
         String createBucketPath = BucketURL.CREATE_BUCKET.getUrl().formatted(name);
         URI uri = URI.create("%s/%s".formatted(reductClient.getServerProperties().getBaseUrl(), createBucketPath));
         HttpRequest.Builder httpRequest = HttpRequest.newBuilder()
                 .uri(uri)
-                .POST(HttpRequest.BodyPublishers.ofString(JsonUtils.serialize(bucketSettings)));
+                .PUT(HttpRequest.BodyPublishers.ofString(JsonUtils.serialize(bucketSettings)));
         reductClient.send(httpRequest, HttpResponse.BodyHandlers.discarding()); //TODO ask about default settings. The answer from DB always is empty for success, but settings sets as default. This bucket will always have settings as null until invoke read.
-        return this;
     }
 
     /**
+     * The method updates the Bucket and returns updated BucketSettings. If authentication is enabled, the method needs a valid API token.
+     * @return BucketSettings
+     */
+    @JsonIgnore
+    public BucketSettings getSettings() {
+        return this.read().getBucketSettings();
+    }
+    /**
      * Write a record to an entry.
-     * @param entry
+     * @param record
      * @throws ReductException
-     * @throws ReductSDKException
+     * @throws ReductException
      * @throws IllegalArgumentException
      */
-    public void writeRecord(@NonNull Entry<?> entry) throws ReductException, ReductSDKException, IllegalArgumentException {
+    public void writeRecord(@NonNull Record record) throws ReductException, IllegalArgumentException {
         //TODO validation block
-        if(Strings.isBlank(entry.getName())
-                || Objects.isNull(entry.getTimestamp())
-                || entry.getTimestamp() <= 0
-                || Objects.isNull(entry.getBody()))
+        if(isNotValidRecord(record))
         {
-            throw new ReductSDKException("Validation error");
+            throw new ReductException("Validation error");
         }
-        String timeStampQuery = getTimestampQuery(entry.getTimestamp());
-        URI uri = URI.create(reductClient.getServerProperties().getBaseUrl() + String.format(EntryURL.WRITE_ENTRY.getUrl(), name, entry.getName()) + timeStampQuery);
+        Long timestamp = Objects.nonNull(record.getTimestamp()) && record.getTimestamp() > 0 ? record.getTimestamp() : Instant.now().getNano()/1000;
+
+        URI uri = URI.create(reductClient.getServerProperties().getBaseUrl() + String.format(RecordURL.WRITE_ENTRY.getUrl(), name, record.getEntryName()) + new Queries(TS, timestamp));
         HttpRequest.Builder builder = HttpRequest.newBuilder()
                 .uri(uri)
-                .header(HttpHeaders.CONTENT_TYPE.getValue(), entry.getBodyClass())
-                .POST(HttpRequest.BodyPublishers.ofByteArray(entry.getByteBodyArray()));
+                .header(getContentTypeHeader(), record.getType())
+                .POST(HttpRequest.BodyPublishers.ofByteArray(record.getBody()));
 
         reductClient.send(builder, HttpResponse.BodyHandlers.ofString()).body();
     }
 
     /**
-     * Get a record from an entry.
-     * @param entry
-     * @return
-     * @param <T>
+     * Write batch of records
+     * @param entryName
+     * @param records
      * @throws ReductException
-     * @throws ReductSDKException
      * @throws IllegalArgumentException
      */
-    public <T extends Serializable> Entry<T> getRecord(Entry<?> entry) throws ReductException, ReductSDKException, IllegalArgumentException {
-        if(Strings.isBlank(name) || Strings.isBlank(entry.getName()))
-        {
-            throw new ReductSDKException("Validation error");
+    public void writeRecords(String entryName, Iterator<Record> records) throws ReductException, IllegalArgumentException {
+        URI uri = URI.create(reductClient.getServerProperties().getBaseUrl() + String.format(RecordURL.WRITE_ENTRY_BATCH.getUrl(), name, entryName));
+        HttpRequest.Builder builder = HttpRequest.newBuilder()
+                .uri(uri);
+
+        byte[] body = null;
+        while (records.hasNext()) {
+            Record record = records.next();
+            //TODO validation block
+            if(isNotValidRecord(record)) {
+                throw new ReductException("Validation error");
+            }
+            byte[] byteBodyArray = record.getBody();
+            body = ArrayUtils.addAll(body, byteBodyArray);
+            builder.header(getXReductTimeWithNumberHeader(record.getTimestamp()), byteBodyArray.length + "," + record.getType());
         }
-        Long timestamp = entry.getTimestamp();
-        String timeStampQuery = Objects.isNull(timestamp) ? "" : getTimestampQuery(timestamp);
-        URI uri = URI.create(reductClient.getServerProperties().getBaseUrl() + String.format(EntryURL.GET_ENTRY.getUrl(), name, entry.getName()) + timeStampQuery);
+        if(Objects.nonNull(body)) {
+            builder.POST(HttpRequest.BodyPublishers.ofByteArray(body));
+            reductClient.send(builder, HttpResponse.BodyHandlers.ofString()).body();
+        }
+    }
+
+    /**
+     * Get a record from an entry.
+     * @param entryName
+     * @param timestamp
+     * @return
+     * @throws ReductException
+     * @throws IllegalArgumentException
+     */
+    public Record readRecord(String entryName, Long timestamp) throws ReductException, IllegalArgumentException {
+        if(Strings.isBlank(name) || Strings.isBlank(entryName))
+        {
+            throw new ReductException("Validation error");
+        }
+        String timeStampQuery = Objects.isNull(timestamp) ? "" : new Queries(TS, timestamp).toString();
+        URI uri = URI.create(reductClient.getServerProperties().getBaseUrl() + String.format(RecordURL.GET_ENTRY.getUrl(), name, entryName) + timeStampQuery);
         HttpRequest.Builder builder = HttpRequest.newBuilder()
                 .uri(uri)
                 .GET();
         HttpResponse<byte[]> httpResponse = reductClient.send(builder, HttpResponse.BodyHandlers.ofByteArray());
-        return (Entry<T>) Entry.builder()
-                .body((T) SerializationUtils.deserialize(httpResponse.body()))
-                .name(entry.getName())
-                .timestamp(httpResponse.headers().firstValue("x-reduct-time").map(Long::getLong).orElse(null))
+        return Record.builder()
+                .body(httpResponse.body())
+                .entryName(entryName)
+                .timestamp(httpResponse.headers().firstValue(getXReductTimeHeader()).map(Long::parseLong).orElseThrow(() -> new ReductException(X_REDUCT_TIME_IS_NOT_SUCH_LONG_FORMAT)))
+                .type(httpResponse.headers().firstValue(getContentTypeHeader()).orElseThrow(() -> new ReductException(CONTENT_TYPE_IS_NOT_SET_IN_THE_RECORD)))
+                .length(httpResponse.headers().firstValue(getContentLengthHeader()).map(Integer::parseInt).orElseThrow(() -> new ReductException(CONTENT_LENGTH_IS_NOT_SET_IN_THE_RECORD)))
                 .build();
     }
 
     /**
      * Get only meta information about record.
-     * @param entry
+     * @param entryName
+     * @param timestamp
      * @return
-     * @param <T>
      * @throws ReductException
-     * @throws ReductSDKException
      * @throws IllegalArgumentException
      */
-    public <T extends Serializable> Entry<T> getMetaInfo(Entry<?> entry) throws ReductException, ReductSDKException, IllegalArgumentException {
-        if(Strings.isBlank(name) || Strings.isBlank(entry.getName()))
+    public Record getMetaInfo(String entryName, Long timestamp) throws ReductException, IllegalArgumentException {
+        if(Strings.isBlank(name) || Strings.isBlank(entryName))
         {
-            throw new ReductSDKException("Validation error");
+            throw new ReductException("Validation error");
         }
-        Long timestamp = entry.getTimestamp();
-        String timeStampQuery = Objects.isNull(timestamp) ? "" : getTimestampQuery(timestamp);
-        URI uri = URI.create(reductClient.getServerProperties().getBaseUrl() + String.format(EntryURL.GET_ENTRY.getUrl(), name, entry.getName()) + timeStampQuery);
+        String timeStampQuery = Objects.isNull(timestamp) ? "" : new Queries(TS, timestamp).toString();
+        URI uri = URI.create(reductClient.getServerProperties().getBaseUrl() + String.format(RecordURL.GET_ENTRY.getUrl(), name, entryName) + timeStampQuery);
         HttpRequest.Builder builder = HttpRequest.newBuilder()
                 .uri(uri)
                 .method("HEAD", HttpRequest.BodyPublishers.noBody());
         HttpResponse<byte[]> httpResponse = reductClient.send(builder, HttpResponse.BodyHandlers.ofByteArray());
-        return (Entry<T>) Entry.builder()
-                .name(entry.getName())
-                .timestamp(httpResponse.headers().firstValue("x-reduct-time").map(Long::valueOf).orElse(null))
+        return Record.builder()
+                .entryName(entryName)
+                .timestamp(httpResponse.headers().firstValue(getXReductTimeHeader()).map(Long::parseLong).orElseThrow(() -> new ReductException(X_REDUCT_TIME_IS_NOT_SUCH_LONG_FORMAT)))
+                .type(httpResponse.headers().firstValue(getContentTypeHeader()).orElseThrow(() -> new ReductException(CONTENT_TYPE_IS_NOT_SET_IN_THE_RECORD)))
+                .length(httpResponse.headers().firstValue(getContentLengthHeader()).map(Integer::parseInt).orElseThrow(() -> new ReductException(CONTENT_LENGTH_IS_NOT_SET_IN_THE_RECORD)))
                 .build();
     }
 
-    private String getTimestampQuery(Long timestamp) {
-        return String.format("?" + TIME_STAMP.getValue() +"=%d", timestamp);
+    /**
+     * Query records for a time interval
+     * @param entryName
+     * @param start
+     * @param stop
+     * @param ttl
+     * @return
+     */
+    public Iterator<Record> query(String entryName, Integer start, Integer stop, Integer ttl) {
+        if(Strings.isBlank(name) || Strings.isBlank(entryName) || Objects.isNull(start) || Objects.isNull(stop) || Objects.isNull(ttl))
+        {
+            throw new ReductException("Validation error");
+        }
+        URI uri = URI.create(reductClient.getServerProperties().getBaseUrl() +
+            String.format(RecordURL.QUERY.getUrl(), name, entryName) + new Queries("start", start).add("stop", stop).add("ttl", ttl));
+        HttpRequest.Builder builder = HttpRequest.newBuilder()
+            .uri(uri)
+            .GET();
+        HttpResponse<String> response = reductClient.send(builder, HttpResponse.BodyHandlers.ofString());
+        QueryId queryId = JsonUtils.parseObject(response.body(), QueryId.class);
+        return getRecords(entryName, queryId.getId()).iterator();
+    }
+
+    /**
+     * Get a bulk of records from an entry
+     * @param entryName
+     * @param queryId
+     * @return
+     * @throws ReductException
+     * @throws IllegalArgumentException
+     */
+    Stream<Record> getRecords(@NonNull String entryName, Long queryId) throws ReductException, IllegalArgumentException {
+        if(Objects.isNull(queryId))
+        {
+            throw new ReductException("Validation error");
+        }
+        URI uri = URI.create(reductClient.getServerProperties().getBaseUrl() + String.format(RecordURL.GET_ENTRIES.getUrl(), name, entryName) + new Queries("q", queryId));
+        HttpRequest.Builder builder = HttpRequest.newBuilder()
+            .uri(uri)
+            .GET();
+        HttpResponse<byte[]> httpResponse = reductClient.send(builder, HttpResponse.BodyHandlers.ofByteArray());
+        ByteBuffer byteBuffer = ByteBuffer.wrap(httpResponse.body());
+        return httpResponse.headers()
+            .map()
+            .entrySet()
+            .stream()
+            .filter(ent -> ent.getKey().contains(getXReductTimeWithUnderscoreHeader()))
+            .map(ent -> {
+                String ts = ent.getKey().substring(getXReductTimeWithUnderscoreHeader().length());
+                String[] split = ent.getValue().get(0).split(",");
+                if(split.length < 2) {
+                    throw new ReductException(String.format("Headers has a wrong format for timestamp: %s", ts));
+                }
+                try {
+                    int length = Integer.parseInt(split[0]);
+                    String type = split[1];
+                    byte[] tempBuf = new byte[length];
+                    byteBuffer.get(tempBuf);
+                    return Record.builder()
+                        .body(tempBuf)
+                        .entryName(entryName)
+                        .timestamp(Optional.of(ts).map(Long::parseLong).orElseThrow(() -> new ReductException(X_REDUCT_TIME_IS_NOT_SUCH_LONG_FORMAT)))
+                        .type(type)
+                        .length(length)
+                        .build();
+                }
+                catch (NumberFormatException ex) {
+                    throw new ReductException(CONTENT_LENGTH_IS_NOT_SET_IN_THE_RECORD);
+                }
+            });
+    }
+
+
+    Collection<Record> getMetaInfos(Record record, Long queryId) throws ReductException, IllegalArgumentException {
+        if(Strings.isBlank(name) || Strings.isBlank(record.getEntryName()) || Objects.isNull(queryId))
+        {
+            throw new ReductException("Validation error");
+        }
+        URI uri = URI.create(reductClient.getServerProperties().getBaseUrl() + String.format(RecordURL.GET_ENTRIES.getUrl(), name, record.getEntryName()) + new Queries("q", queryId));
+        HttpRequest.Builder builder = HttpRequest.newBuilder()
+            .uri(uri)
+            .method("HEAD", HttpRequest.BodyPublishers.noBody());
+        HttpResponse<byte[]> httpResponse = reductClient.send(builder, HttpResponse.BodyHandlers.ofByteArray());
+        return httpResponse.headers()
+            .map()
+            .entrySet()
+            .stream()
+            .filter(ent -> ent.getKey().contains(getXReductTimeWithUnderscoreHeader()))
+            .map(ent -> {
+                String ts = ent.getKey().substring(getXReductTimeWithUnderscoreHeader().length());
+                String[] split = ent.getValue().get(0).split(",");
+                if(split.length < 2) {
+                    throw new ReductException(String.format("Headers has a wrong format for timestamp: %s", ts));
+                }
+                try {
+                    int length = Integer.parseInt(split[0]);
+                    String type = split[1];
+                    byte[] tempBuf = new byte[length];
+                    return Record.builder()
+                        .body(tempBuf)
+                        .entryName(record.getEntryName())
+                        .timestamp(Optional.of(ts).map(Long::parseLong).orElseThrow(() -> new ReductException(X_REDUCT_TIME_IS_NOT_SUCH_LONG_FORMAT)))
+                        .type(type)
+                        .length(length)
+                        .build();
+                }
+                catch (NumberFormatException ex) {
+                    throw new ReductException(CONTENT_LENGTH_IS_NOT_SET_IN_THE_RECORD);
+                }
+            })
+            .collect(Collectors.toCollection(LinkedList::new));
+    }
+
+    private boolean isNotValidRecord(Record record) {
+        return Strings.isBlank(record.getEntryName())
+                || record.getTimestamp() <= 0
+                || Objects.isNull(record.getBody());
     }
 }