+++ src/main/java/org/reduct/utils/JsonUtils.java
@@ -4,12 +4,11 @@ import com.fasterxml.jackson.core.JacksonException;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.experimental.UtilityClass;
-import org.reduct.common.exception.ReductSDKException;
+import org.reduct.common.exception.ReductException;
 import org.reduct.model.bucket.BucketSettings;
 import org.reduct.model.token.TokenPermissions;
 
 import java.util.List;
-import java.util.Objects;
 
 @UtilityClass
 public class JsonUtils {
@@ -18,14 +17,14 @@ public class JsonUtils {
         try {
             return objectMapper.readValue(jsonBody, tClass);
         } catch (JacksonException e) {
-            throw new ReductSDKException("The server returned a malformed response.");
+            throw new ReductException(e.getMessage());
         }
     }
     public static <T> List<T> parseObjectAsList(String jsonBody, Class<T> tClass) {
         try {
             return objectMapper.readValue(jsonBody, objectMapper.getTypeFactory().constructCollectionType(List.class, tClass));
         } catch (JacksonException e) {
-            throw new ReductSDKException("The server returned a malformed response.");
+            throw new ReductException("The server returned a malformed response.");
         }
     }
 
@@ -36,7 +35,7 @@ public class JsonUtils {
         try {
             return objectMapper.writeValueAsString(bucketSettings);
         } catch (JsonProcessingException e) {
-            throw new ReductSDKException("Failed to serialize bucket settings", e);
+            throw new ReductException("Failed to serialize bucket settings", e);
         }
     }
 
@@ -44,7 +43,7 @@ public class JsonUtils {
         try {
             return objectMapper.writeValueAsString(permissions);
         } catch (JsonProcessingException e) {
-            throw new ReductSDKException("Failed to serialize the token permissions.", e);
+            throw new ReductException("Failed to serialize the token permissions.", e);
         }
     }
 
@@ -52,7 +51,7 @@ public class JsonUtils {
         try {
             return objectMapper.writeValueAsString(obj);
         } catch (JsonProcessingException e) {
-            throw new ReductSDKException("Failed to serialize.", e);
+            throw new ReductException("Failed to serialize.", e);
         }
     }
 }